requirements.txt
aiofiles==23.1.0
aiogram==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
automium==0.2.6
automium-web==0.1.1
blinker==1.7.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
distro==1.8.0
Flask==3.0.0
frozenlist==1.4.0
google-api-core==2.12.0
google-auth==2.23.3
google-cloud-vision==3.4.4
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-status==1.59.0
gTTS==2.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
legacy==0.1.7
magic-filter==1.0.11
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.26.0
openai==1.3.3
Pillow==10.1.0
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
sniffio==1.3.0
sounddevice==0.4.6
SpeechRecognition==3.10.0
SQLAlchemy==2.0.23
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==3.0.1
yarl==1.9.2
